refactor(admin): split adminClient.ts into 6 domain sub-clients
Split the 3878-line adminClient.ts into 6 domain-specific sub-client
files under clients/admin/, keeping a thin factory in adminClient.ts
that composes them via spread for backward compatibility.

Sub-clients:
- adminUsersClient.ts: User CRUD, sessions, impersonation, API keys
- adminContentClient.ts: Prompts, audit logs, shared games, entity links, email
- adminAiClient.ts: AI models, LLM config, tier/strategy, A/B testing, RAG/KB
- adminAnalyticsClient.ts: Stats, usage, tokens, cost, forecasting, ledger
- adminSystemClient.ts: Workflows, N8n, infrastructure, batch jobs, queue
- adminMonitorClient.ts: Service dashboard, metrics, reports, OpenRouter

No consumer files need changes - the factory pattern preserves the
existing api.admin.methodName() interface. All 85 admin tests pass.
