+++ apps/web/src/lib/api/clients/admin/index.ts
@@ -1,0 +1,6 @@
+export { createAdminUsersClient, type AdminUsersClient } from './adminUsersClient';
+export { createAdminContentClient, type AdminContentClient } from './adminContentClient';
+export { createAdminAiClient, type AdminAiClient } from './adminAiClient';
+export { createAdminAnalyticsClient, type AdminAnalyticsClient } from './adminAnalyticsClient';
+export { createAdminSystemClient, type AdminSystemClient } from './adminSystemClient';
+export { createAdminMonitorClient, type AdminMonitorClient } from './adminMonitorClient';
